A bit more
So I continued a bit and had a discussion with another Unity programmer. We agreed that Unity UI is horrible for vector math, and I chose to change it up a bit. I also found a PICNIC, so changed that, and now position works perfectly. Just the rotation and scale to go now.