Delay and Test
I've added the delay function! Yay! It looks cool and works really well. I've also started the work on a test case, where an agent is collecting data while I move it around, which is then sent to the ANN handler to create a new ANN. I just need to fix a problem with how outputs are handled. I got the desired outputs in a list, I just need to save them into the output neurons and NOT make a neuron for each! The data is meant for comparison on each run through the data, and there should only be created an output neuron for each output (agent movement input)!