Visual Update
I changed how the visualization is performed. With this update, the visual connection to the actual Neuron class has been removed and replaced by a delegate event that the visual neurons can subscribe to instead.
This update is based on an error of trying to update the visuals from a sub-thread, which isn't allowed, so this is an attempt at fixing a major problem of ANN and visual being too intertwined.