Fixes & Saving
I found a major error in the backpropagation. I was using the derivative of the output layer AF for every hidden layer calculation... God dammit.
I also sorted some stuff in the assets into new folders and sub-folders (SO stuff).
Lastly, I made it so I can save and load ANN's using SO's.